Module canyon_cli::chain_spec

Substrate chain configurations.

Structs

Extensions
Node ChainSpec extensions.

GenesisConfig

Functions

authority_keys_from_seed
Helper function to generate stash, controller and session key from seed

development_config
Development config (single validator Alice)

get_account_id_from_seed
Helper function to generate an account ID from seed

get_from_seed
Helper function to generate a crypto pair from seed

local_testnet_config
Local testnet config (multivalidator Alice + Bob)

staging_testnet_config
Staging testnet config.

testnet_genesis
Helper function to create GenesisConfig for testing

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance
Balance of an account.

ChainSpec
Specialized ChainSpec.

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.